added music service abstract class for adding more services
